requirements.txt
fastapi
uvicorn
pandas
joblib
catboost
scikit-learn
requests
beautifulsoup4
tqdm
python-dotenv
schedule
openpyxl
holidays
sqlalchemy
psycopg[binary,pool]